// apps/webapp/spa/src/lib/api/generated/models/BrandAudienceAgeRange.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Voz AI API
 * Voz AI marketing automation platform API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const BrandAudienceAgeRange = {
    teens: 'TEENS',
    youngAdults: 'YOUNG_ADULTS',
    adults: 'ADULTS',
    middleAged: 'MIDDLE_AGED',
    seniors: 'SENIORS',
    any: 'ANY'
} as const;
export type BrandAudienceAgeRange = typeof BrandAudienceAgeRange[keyof typeof BrandAudienceAgeRange];


export function instanceOfBrandAudienceAgeRange(value: any): boolean {
    for (const key in BrandAudienceAgeRange) {
        if (Object.prototype.hasOwnProperty.call(BrandAudienceAgeRange, key)) {
            if (BrandAudienceAgeRange[key as keyof typeof BrandAudienceAgeRange] === value) {
                return true;
            }
        }
    }
    return false;
}

export function BrandAudienceAgeRangeFromJSON(json: any): BrandAudienceAgeRange {
    return BrandAudienceAgeRangeFromJSONTyped(json, false);
}

export function BrandAudienceAgeRangeFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrandAudienceAgeRange {
    return json as BrandAudienceAgeRange;
}

export function BrandAudienceAgeRangeToJSON(value?: BrandAudienceAgeRange | null): any {
    return value as any;
}

export function BrandAudienceAgeRangeToJSONTyped(value: any, ignoreDiscriminator: boolean): BrandAudienceAgeRange {
    return value as BrandAudienceAgeRange;
}