// apps/webapp/spa/src/lib/api/generated/models/BrandAudienceGender.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Voz AI API
 * Voz AI marketing automation platform API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const BrandAudienceGender = {
    male: 'MALE',
    female: 'FEMALE',
    any: 'ANY'
} as const;
export type BrandAudienceGender = typeof BrandAudienceGender[keyof typeof BrandAudienceGender];


export function instanceOfBrandAudienceGender(value: any): boolean {
    for (const key in BrandAudienceGender) {
        if (Object.prototype.hasOwnProperty.call(BrandAudienceGender, key)) {
            if (BrandAudienceGender[key as keyof typeof BrandAudienceGender] === value) {
                return true;
            }
        }
    }
    return false;
}

export function BrandAudienceGenderFromJSON(json: any): BrandAudienceGender {
    return BrandAudienceGenderFromJSONTyped(json, false);
}

export function BrandAudienceGenderFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrandAudienceGender {
    return json as BrandAudienceGender;
}

export function BrandAudienceGenderToJSON(value?: BrandAudienceGender | null): any {
    return value as any;
}

export function BrandAudienceGenderToJSONTyped(value: any, ignoreDiscriminator: boolean): BrandAudienceGender {
    return value as BrandAudienceGender;
}