// apps/webapp/spa/src/lib/api/generated/models/BrandAudienceIncomeRange.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Voz AI API
 * Voz AI marketing automation platform API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const BrandAudienceIncomeRange = {
    lowIncome: 'LOW_INCOME',
    middleIncome: 'MIDDLE_INCOME',
    upperMiddleIncome: 'UPPER_MIDDLE_INCOME',
    highIncome: 'HIGH_INCOME',
    any: 'ANY'
} as const;
export type BrandAudienceIncomeRange = typeof BrandAudienceIncomeRange[keyof typeof BrandAudienceIncomeRange];


export function instanceOfBrandAudienceIncomeRange(value: any): boolean {
    for (const key in BrandAudienceIncomeRange) {
        if (Object.prototype.hasOwnProperty.call(BrandAudienceIncomeRange, key)) {
            if (BrandAudienceIncomeRange[key as keyof typeof BrandAudienceIncomeRange] === value) {
                return true;
            }
        }
    }
    return false;
}

export function BrandAudienceIncomeRangeFromJSON(json: any): BrandAudienceIncomeRange {
    return BrandAudienceIncomeRangeFromJSONTyped(json, false);
}

export function BrandAudienceIncomeRangeFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrandAudienceIncomeRange {
    return json as BrandAudienceIncomeRange;
}

export function BrandAudienceIncomeRangeToJSON(value?: BrandAudienceIncomeRange | null): any {
    return value as any;
}

export function BrandAudienceIncomeRangeToJSONTyped(value: any, ignoreDiscriminator: boolean): BrandAudienceIncomeRange {
    return value as BrandAudienceIncomeRange;
}