// apps/webapp/spa/src/lib/api/generated/models/ContentChannelName.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Voz AI API
 * Voz AI marketing automation platform API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ContentChannelName = {
    instagram: 'INSTAGRAM',
    linkedin: 'LINKEDIN'
} as const;
export type ContentChannelName = typeof ContentChannelName[keyof typeof ContentChannelName];


export function instanceOfContentChannelName(value: any): boolean {
    for (const key in ContentChannelName) {
        if (Object.prototype.hasOwnProperty.call(ContentChannelName, key)) {
            if (ContentChannelName[key as keyof typeof ContentChannelName] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ContentChannelNameFromJSON(json: any): ContentChannelName {
    return ContentChannelNameFromJSONTyped(json, false);
}

export function ContentChannelNameFromJSONTyped(json: any, ignoreDiscriminator: boolean): ContentChannelName {
    return json as ContentChannelName;
}

export function ContentChannelNameToJSON(value?: ContentChannelName | null): any {
    return value as any;
}

export function ContentChannelNameToJSONTyped(value: any, ignoreDiscriminator: boolean): ContentChannelName {
    return value as ContentChannelName;
}